requirements.txt
aiofiles==24.1.0
aiogram==3.19.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
asyncpg==0.30.0
attrs==25.3.0
certifi==2025.1.31
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
magic-filter==1.0.12
multidict==6.2.0
numpy==2.2.4
pandas==2.2.3
propcache==0.3.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.40
typing_extensions==4.13.0
tzdata==2025.2
XlsxWriter==3.2.2
yarl==1.18.3
